✨feat: 대시보드 지수현황,지수차트,지수 성과 분석 랭킹

+++ src/main/java/com/ssibssaggi/findex/controller/dto/IndexDataResponse.java
@@ -19,8 +19,8 @@ public record IndexDataResponse(
         Float versus,
         Float fluctuationRate,
         Long tradingQuantity,
-        Float tradingPrice,
-        Float marketTotalAmount
+        Long tradingPrice,
+        Long marketTotalAmount
 ) {
 
     // IndexData 엔티티 -> IndexDataResponse 변환
